+++ tests/testthat/test-connections.R
@@ -1,13 +1,58 @@
 context("catmaid login and get/post")
 
+# we can only run real tests if we can log in with default parameters
+conn=try(catmaid_login(Force = TRUE), silent = TRUE)
+
+# Sample server used in next block
+publicserver <- 'http://hildebrand16.neurodata.io/catmaid/' #public server running a catmaid instance..
+
+test_that("set and get the environmental variables responsible for connections", {
+
+  #first check if any environment variable has the name of format catmaid. or catmaid_
+  catmaid_msg <- try(catmaid_envstr())
+  if (class(catmaid_msg)[[1]] == 'simpleMessage'){
+    expect_message(message(catmaid_msg), regexp = "No usable environment")
+    skip('No environmental variables found so skipping..')
+  } else if (class(catmaid_msg) == "try-error"){
+    #some environmental variables are missing here, so best to skip the tests again..
+    skip('Some environmental variables are missing so skipping..')
+  }
+
+  # check the named object has valid field names
+  expect_true(all(
+    names(catmaid_connection_getenv()) %in%
+      c(
+        "server",
+        "authname",
+        "authpassword",
+        "authtype",
+        "token",
+        "username",
+        "password"
+      )
+  ))
+  
+  #Set environment variables through connection object and check them back again..
+  pubconn <- catmaid_connection(server=publicserver, authname = "flyeee")
+  catmaid_connection_setenv(conn = pubconn, Cache=FALSE)
+  expect_match(catmaid_connection_getenv()['server'],publicserver)
+  expect_match(catmaid_connection_getenv()['authname'],'flyeee')
+  
+  # reset environment vars so that other test cases can run
+  # we reset them to those implied by the default connection if it is valid
+  # otherwise we just unset them
+  if(inherits(conn, 'try-error'))
+    catmaid_connection_unsetenv()
+  else
+    catmaid_connection_setenv(conn = conn)
+})
+
 test_that("can connect to a server without logging in", {
   # No login is required to talk to this server
   expect_is(catmaid_login(server='http://hildebrand16.neurodata.io/catmaid/', Cache = FALSE),
             'catmaid_connection')
 })
 
-# we can only run real tests if we can log in with default parameters
-conn=try(catmaid_login(Force = TRUE), silent = TRUE)
 
 test_that("can make a connection", {
   
@@ -38,7 +83,7 @@ test_that("can get and post data", {
   expect_is(skel<-catmaid_fetch("1/10418394/0/0/compact-skeleton", conn=conn, parse.json = FALSE),
             'HttpResponse')
   expect_is(neuronnames<-catmaid_fetch("/1/skeleton/neuronnames", conn=conn,
-                                  body=list(pid=1, 'skids[1]'=10418394, 'skids[2]'=4453485)),
+                                       body=list(pid=1, 'skids[1]'=10418394, 'skids[2]'=4453485)),
             'list')
   # nb we can't rely on the returned order
   expect_equal(sort(names(neuronnames)), c('10418394','4453485'))